fix: mention options without function
