+++ packages/mentis/src/utils/cn.ts
@@ -1,3 +1,3 @@
-export const cn = (...classes: (string | undefined)[]) => {
+export const cn = (...classes: (string | undefined)[]): string => {
   return classes.filter(Boolean).join(" ");
 };
