+++ packages/mentis/src/utils/filterOutOptionFunctions.ts
@@ -1,0 +1,22 @@
+import type { MentionOption } from "../types/MentionInput.types";
+
+export type MentionOptionWithoutFunction = Omit<MentionOption, "value"> & {
+  value: string;
+};
+
+export const filterOutOptionFunctions = (
+  options: MentionOption[]
+): MentionOptionWithoutFunction[] => {
+  const optionsWithoutFunctions: MentionOptionWithoutFunction[] = [];
+
+  options.forEach((option) => {
+    if (typeof option.value !== "function") {
+      optionsWithoutFunctions.push({
+        ...option,
+        value: option.value,
+      });
+    }
+  });
+
+  return optionsWithoutFunctions;
+};
